feat(profiling) lazy init ui frames renderer
