+++ static/app/utils/profiling/renderers/uiFramesRendererWebGL.tsx
@@ -1,5 +1,5 @@
 import * as Sentry from '@sentry/react';
-import {mat3, vec2} from 'gl-matrix';
+import {mat3} from 'gl-matrix';
 
 import {FlamegraphTheme} from 'sentry/utils/profiling/flamegraph/flamegraphTheme';
 import {
@@ -12,11 +12,10 @@ import {
   pointToAndEnableVertexAttribute,
   resizeCanvasToDisplaySize,
   safeGetContext,
-  upperBound,
 } from 'sentry/utils/profiling/gl/utils';
 import {UIFramesRenderer} from 'sentry/utils/profiling/renderers/UIFramesRenderer';
 import {Rect} from 'sentry/utils/profiling/speedscope';
-import {UIFrameNode, UIFrames} from 'sentry/utils/profiling/uiFrames';
+import {UIFrames} from 'sentry/utils/profiling/uiFrames';
 
 import {uiFramesFragment, uiFramesVertext} from './shaders';
 
@@ -237,40 +236,6 @@ class UIFramesRendererWebGL extends UIFramesRenderer {
       return this.theme.COLORS.UI_FRAME_COLOR_SLOW;
     }
     throw new Error(`Invalid frame type - ${type}`);
-  }
-
-  findHoveredNode(configSpaceCursor: vec2, configSpace: Rect): UIFrameNode[] | null {
-    // ConfigSpace origin is at top of rectangle, so we need to offset bottom by 1
-    // to account for size of renderered rectangle.
-    if (configSpaceCursor[1] > configSpace.bottom + 1) {
-      return null;
-    }
-
-    if (configSpaceCursor[0] < configSpace.left) {
-      return null;
-    }
-
-    if (configSpaceCursor[0] > configSpace.right) {
-      return null;
-    }
-
-    const overlaps: UIFrameNode[] = [];
-    // We can find the upper boundary, but because frames might overlap, we need to also check anything
-    // before the upper boundary to see if it overlaps... Performance does not seem to be a big concern
-    // here as the max number of slow frames we can have is max profile duration / slow frame = 30000/
-    const end = upperBound(configSpaceCursor[0], this.uiFrames.frames);
-
-    for (let i = 0; i < end; i++) {
-      const frame = this.uiFrames.frames[i];
-      if (configSpaceCursor[0] <= frame.end && configSpaceCursor[0] >= frame.start) {
-        overlaps.push(frame);
-      }
-    }
-
-    if (overlaps.length > 0) {
-      return overlaps;
-    }
-    return null;
   }
 
   draw(configViewToPhysicalSpace: mat3): void {
